+++ test/test_interfaces.py
@@ -30,7 +30,10 @@ class ModbusInterfaceTestsTest(unittest.TestCase):
         x = None
         instance = IModbusDecoder()
         self.assertRaises(NotImplementedException, lambda: instance.decode(x))
-        self.assertRaises(NotImplementedException, lambda: instance.lookupPduClass(x))
+        self.assertRaises(NotImplementedException,
+                          lambda: instance.lookupPduClass(x))
+        self.assertRaises(NotImplementedException,
+                          lambda: instance.register(x))
 
     def testModbusFramerInterface(self):
         ''' Test that the base class isn't implemented '''