+++ pymodbus/framer/rtu_framer.py
@@ -205,9 +205,10 @@ class ModbusRtuFramer(ModbusFramer):
 
         :param data: The new packet data
         :param callback: The function to send results to
-        :param unit: Process if unit id matches, ignore otherwise (could be a \
-        list of unit ids (server) or single unit id(client/server)
+        :param unit: Process if unit id matches, ignore otherwise (could be a
+               list of unit ids (server) or single unit id(client/server)
         :param single: True or False (If True, ignore unit address validation)
+
         """
         if not isinstance(unit, (list, tuple)):
             unit = [unit]